ref: move current ResultsPage to StorageResponsesTab component
